LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/manual/wide-gamut-canvas/canvas-drawImage-e_srgb-expected.txt


FAIL Draw SRGB image on an e-sRGB canvas and read back the e-sRGB pixels. assert_approx_equals: expected 0.88623 +/- 0.01 but got 225
